requirements.txt
django
gunicorn
whitenoise
dj-database-url
python-dotenv
psycopg2-binary
mysqlclient
django-allauth
requests
beautifulsoup4
bs4 